# ClearBits with AndNot