I am not entirely sure why we had the iterator set up.

Clearing bits with AndNot feels correct.

Let me know if there's some case that I've missed here.

Did not remove the `ClearBits` functions as they're exported public API.