@guymolinari 2025-08-07
The intention of ClearBits() is to perform "bulk deletes" of values.

El jue, 7 de ago de 2025, 12:50 a.m., Piyush ***@***.***>
escribió:

> I am not entirely sure why we had the iterator set up.
>
> Clearing bits with AndNot feels correct.
>
> Let me know if there's some case that I've missed here.
>
> Did not remove the ClearBits functions as they're exported public API.
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/RoaringBitmap/roaring/pull/480
> Commit Summary
>
>    - d4f5fa0
>    <https://github.com/RoaringBitmap/roaring/pull/480/commits/d4f5fa0307c5d63e844118bd620495bbc4c5a35f>
>    clear bits with and not
>
> File Changes
>
> (2 files <https://github.com/RoaringBitmap/roaring/pull/480/files>)
>
>    - *M* BitSliceIndexing/bsi.go
>    <https://github.com/RoaringBitmap/roaring/pull/480/files#diff-6c94533959c140be49bad4001c54efb949159a7c63b23a72891a61e00be17091>
>    (10)
>    - *M* roaring64/bsi64.go
>    <https://github.com/RoaringBitmap/roaring/pull/480/files#diff-d1a31b49d0996408823b8c849e138363be8a4347d2e57cafd0ad289768845b46>
>    (12)
>
> Patch Links:
>
>    - https://github.com/RoaringBitmap/roaring/pull/480.patch
>    - https://github.com/RoaringBitmap/roaring/pull/480.diff
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/RoaringBitmap/roaring/pull/480>, or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ADZZAUUGH7ZEJ4DJOL5DX233MLZNDAVCNFSM6AAAAACDKA6SQSVHI2DSMVQWIX3LMV43ASLTON2WKOZTGI4TSMRTGAYTINQ>
> .
> You are receiving this because you are subscribed to this thread.Message
> ID: ***@***.***>
>