@ofpiyush 2025-08-07
> The intention of ClearBits() is to perform "bulk deletes" of values.

Did you mean `ClearValues` there?

Wouldn't `AndNot` do that and perform better than the iteration?