@guymolinari 2025-08-07
I would answer yes to both questions.

If AndNot performs better then it is an easy decision.   In general, if you
have a performance improvement backed up by a test case that illustrates
the improvement,  a merge is almost guaranteed ;-)

El jue, 7 de ago de 2025, 6:27 a.m., Piyush ***@***.***>
escribió:

> *ofpiyush* left a comment (RoaringBitmap/roaring#480)
> <https://github.com/RoaringBitmap/roaring/pull/480#issuecomment-3163976819>
>
> The intention of ClearBits() is to perform "bulk deletes" of values.
>
> Did you mean ClearValues there?
>
> Wouldn't AndNot do that and perform better than the iteration?
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/RoaringBitmap/roaring/pull/480#issuecomment-3163976819>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ADZZAUT53X33RDUZEE2GJOD3MNA45AVCNFSM6AAAAACDKA6SQSVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZTCNRTHE3TMOBRHE>
> .
> You are receiving this because you commented.Message ID:
> ***@***.***>
>