@ofpiyush 2025-08-07
Benchmark results on my machine. Given the nature of the change, it should translate to linux servers as well.

```
goos: darwin
goarch: arm64
pkg: github.com/RoaringBitmap/roaring/v2/BitSliceIndexing
cpu: Apple M4 Max
               │ bench_old.txt │            bench_new.txt             │
               │    sec/op     │    sec/op     vs base                │
ClearValues-14    18.224m ± 2%   3.968m ± 23%  -78.23% (p=0.000 n=10)

               │ bench_old.txt │            bench_new.txt             │
               │     B/op      │     B/op      vs base                │
ClearValues-14    9.536Ki ± 1%   8.536Ki ± 0%  -10.49% (p=0.000 n=10)

               │ bench_old.txt │           bench_new.txt            │
               │   allocs/op   │ allocs/op   vs base                │
ClearValues-14      29.00 ± 0%   20.00 ± 0%  -31.03% (p=0.000 n=10)
```

```
goos: darwin
goarch: arm64
pkg: github.com/RoaringBitmap/roaring/v2/roaring64
cpu: Apple M4 Max
               │ bench_old.txt │            bench_new.txt            │
               │    sec/op     │   sec/op     vs base                │
ClearValues-14    24.845m ± 2%   5.019m ± 1%  -79.80% (p=0.000 n=10)

               │ bench_old.txt │            bench_new.txt            │
               │     B/op      │     B/op      vs base               │
ClearValues-14    25.54Ki ± 1%   24.55Ki ± 0%  -3.85% (p=0.000 n=10)

               │ bench_old.txt │           bench_new.txt            │
               │   allocs/op   │ allocs/op   vs base                │
ClearValues-14      33.00 ± 0%   24.00 ± 0%  -27.27% (p=0.000 n=10)
```

<details>
<summary>Commands and raw outputs </summary>
<p>

With old code calling ClearBits with the iterator
```
go test -bench='^BenchmarkClearValues$' -run='^$' -benchmem -count=10 > bench_old.txt
```

With `AndNot`
```
go test -bench='^BenchmarkClearValues$' -run='^$' -benchmem -count=10 > bench_new.txt
```

then benchstat
```
benchstat bench_old.txt bench_new.txt
```

Raw data:
```
cat BitSliceIndexing/bench_old.txt

goos: darwin
goarch: arm64
pkg: github.com/RoaringBitmap/roaring/v2/BitSliceIndexing
cpu: Apple M4 Max
BenchmarkClearValues-14    	      66	  17949452 ns/op	   10301 B/op	      30 allocs/op
BenchmarkClearValues-14    	      63	  17658084 ns/op	    9762 B/op	      29 allocs/op
BenchmarkClearValues-14    	      67	  17995859 ns/op	    9768 B/op	      29 allocs/op
BenchmarkClearValues-14    	      67	  17879762 ns/op	    9767 B/op	      29 allocs/op
BenchmarkClearValues-14    	      66	  18207193 ns/op	    9762 B/op	      29 allocs/op
BenchmarkClearValues-14    	      66	  18267813 ns/op	    9756 B/op	      29 allocs/op
BenchmarkClearValues-14    	      63	  18239941 ns/op	    9756 B/op	      29 allocs/op
BenchmarkClearValues-14    	      66	  18322739 ns/op	    9763 B/op	      29 allocs/op
BenchmarkClearValues-14    	      66	  18874647 ns/op	    9770 B/op	      29 allocs/op
BenchmarkClearValues-14    	      64	  18279068 ns/op	    9828 B/op	      29 allocs/op
PASS
ok  	github.com/RoaringBitmap/roaring/v2/BitSliceIndexing	40.108s
```

```
cat BitSliceIndexing/bench_new.txt

goos: darwin
goarch: arm64
pkg: github.com/RoaringBitmap/roaring/v2/BitSliceIndexing
cpu: Apple M4 Max
BenchmarkClearValues-14    	     310	   3849987 ns/op	    8760 B/op	      20 allocs/op
BenchmarkClearValues-14    	     266	   3964120 ns/op	    8763 B/op	      20 allocs/op
BenchmarkClearValues-14    	     226	   4588439 ns/op	    8729 B/op	      20 allocs/op
BenchmarkClearValues-14    	     300	   4688459 ns/op	    8730 B/op	      20 allocs/op
BenchmarkClearValues-14    	     307	   5119839 ns/op	    8740 B/op	      20 allocs/op
BenchmarkClearValues-14    	     306	   3971695 ns/op	    8756 B/op	      20 allocs/op
BenchmarkClearValues-14    	     267	   3880223 ns/op	    8741 B/op	      20 allocs/op
BenchmarkClearValues-14    	     259	   3918312 ns/op	    8766 B/op	      20 allocs/op
BenchmarkClearValues-14    	     226	   4866790 ns/op	    8733 B/op	      20 allocs/op
BenchmarkClearValues-14    	     309	   3942959 ns/op	    8741 B/op	      20 allocs/op
PASS
ok  	github.com/RoaringBitmap/roaring/v2/BitSliceIndexing	145.173s
```

```
cat roaring64/bench_old.txt

goos: darwin
goarch: arm64
pkg: github.com/RoaringBitmap/roaring/v2/roaring64
cpu: Apple M4 Max
BenchmarkClearValues-14    	      48	  24764023 ns/op	   26910 B/op	      34 allocs/op
BenchmarkClearValues-14    	      49	  24731279 ns/op	   26526 B/op	      33 allocs/op
BenchmarkClearValues-14    	      49	  24926856 ns/op	   26152 B/op	      33 allocs/op
BenchmarkClearValues-14    	      48	  24809299 ns/op	   26152 B/op	      33 allocs/op
BenchmarkClearValues-14    	      49	  24738934 ns/op	   26152 B/op	      33 allocs/op
BenchmarkClearValues-14    	      49	  24838062 ns/op	   26152 B/op	      33 allocs/op
BenchmarkClearValues-14    	      49	  24850997 ns/op	   26280 B/op	      33 allocs/op
BenchmarkClearValues-14    	      48	  25561292 ns/op	   26152 B/op	      33 allocs/op
BenchmarkClearValues-14    	      48	  24999984 ns/op	   26152 B/op	      33 allocs/op
BenchmarkClearValues-14    	      48	  25454627 ns/op	   26152 B/op	      33 allocs/op
PASS
ok  	github.com/RoaringBitmap/roaring/v2/roaring64	29.255s
```

```
cat roaring64/bench_new.txt

goos: darwin
goarch: arm64
pkg: github.com/RoaringBitmap/roaring/v2/roaring64
cpu: Apple M4 Max
BenchmarkClearValues-14    	     249	   4945412 ns/op	   25144 B/op	      24 allocs/op
BenchmarkClearValues-14    	     243	   4952487 ns/op	   25162 B/op	      24 allocs/op
BenchmarkClearValues-14    	     235	   5081436 ns/op	   25144 B/op	      24 allocs/op
BenchmarkClearValues-14    	     228	   5017723 ns/op	   25144 B/op	      24 allocs/op
BenchmarkClearValues-14    	     237	   5150613 ns/op	   25151 B/op	      24 allocs/op
BenchmarkClearValues-14    	     238	   5021206 ns/op	   25144 B/op	      24 allocs/op
BenchmarkClearValues-14    	     237	   5026184 ns/op	   25144 B/op	      24 allocs/op
BenchmarkClearValues-14    	     238	   5079279 ns/op	   25144 B/op	      24 allocs/op
BenchmarkClearValues-14    	     238	   5000500 ns/op	   25144 B/op	      24 allocs/op
BenchmarkClearValues-14    	     238	   4998973 ns/op	   25144 B/op	      24 allocs/op
PASS
ok  	github.com/RoaringBitmap/roaring/v2/roaring64	67.009s
```

</p>
</details>